refactor: Update user deletion logging message
The UserController and UserService files were modified to update the logging message when a user is deleted. The previous message included the user ID and the total number of documents deleted, but it has been changed to include the user's email address and ID instead. This change improves the clarity and usefulness of the log message.